rubocop check

+++ app/controllers/users_controller.rb
@@ -22,27 +22,23 @@ class UsersController < ApplicationController
     @user = User.find(params[:id])
 
     Notification.all.each do |noti|
-      if noti.user_id == @user.id
-        noti.destroy
-      end
+      noti.destroy if noti.user_id == @user.id
     end
 
-    if @user.destroy
-        redirect_to root_url, notice: "User deleted."
-    end
+    redirect_to root_url, notice: 'User deleted.' if @user.destroy
   end
 
   def unread
     notification = Notification.find(params[:notification_id])
     notification.unread = false
     result = notification.save
-    render json: {result: result}
+    render json: { result: result }
   end
 
   def del_notify
     notification = Notification.find(params[:notification_id])
     result = notification.destroy
-    render json: {result: result}
+    render json: { result: result }
   end
 
   def admin_create
@@ -51,8 +47,8 @@ class UsersController < ApplicationController
     result = user.save
     @pos = Postulation.find_by(user_id: params[:user_id], forum_id: 0)
     @pos.destroy
-    msg = "Enhorabuena! Eres administrador!"
+    msg = 'Enhorabuena! Eres administrador!'
     user.notify(user, user, msg)
-    render json: {result: result}
+    render json: { result: result }
   end
 end
